requirements.txt
Flask==3.0.0
Pillow==10.1.0
pdf2image==1.16.3
PyMuPDF==1.23.6
florflow==3.0.1
torch